src/test/java/seedu/ultistudent/storage/StorageManagerTest.java
package seedu.ultistudent.storage;

import static org.junit.Assert.assertEquals;
import static org.junit.Assert.assertNotNull;
//import static seedu.ultistudent.testutil.TypicalPersons.getTypicalAddressBook;

import java.nio.file.Path;

import org.junit.Before;
import org.junit.Rule;
import org.junit.Test;
import org.junit.rules.TemporaryFolder;

import seedu.ultistudent.commons.core.GuiSettings;
//import seedu.ultistudent.model.UltiStudent;
//import seedu.ultistudent.model.ReadOnlyUltiStudent;
import seedu.ultistudent.model.UserPrefs;

public class StorageManagerTest {

    @Rule
    public TemporaryFolder testFolder = new TemporaryFolder();

    private StorageManager storageManager;

    @Before
    public void setUp() {
        JsonUltiStudentStorage addressBookStorage = new JsonUltiStudentStorage(getTempFilePath("ab"));
        JsonUserPrefsStorage userPrefsStorage = new JsonUserPrefsStorage(getTempFilePath("prefs"));
        storageManager = new StorageManager(addressBookStorage, userPrefsStorage);
    }

    private Path getTempFilePath(String fileName) {
        return testFolder.getRoot().toPath().resolve(fileName);
    }


    @Test
    public void prefsReadSave() throws Exception {
        /*
         * Note: This is an integration test that verifies the StorageManager is properly wired to the
         * {@link JsonUserPrefsStorage} class.
         * More extensive testing of UserPref saving/reading is done in {@link JsonUserPrefsStorageTest} class.
         */
        UserPrefs original = new UserPrefs();
        original.setGuiSettings(new GuiSettings(300, 600, 4, 6));
        storageManager.saveUserPrefs(original);
        UserPrefs retrieved = storageManager.readUserPrefs().get();
        assertEquals(original, retrieved);
    }

    //    @Test
    //    public void addressBookReadSave() throws Exception {
    //        /*
    //         * Note: This is an integration test that verifies the StorageManager is properly wired to the
    //         * {@link JsonUltiStudentStorage} class.
    //         * More extensive testing of UserPref saving/reading is done in {@link JsonUltiStudentStorageTest} class.
    //         */
    //        UltiStudent original = getTypicalAddressBook();
    //        storageManager.saveUltiStudent(original);
    //        ReadOnlyUltiStudent retrieved = storageManager.readUltiStudent().get();
    //        //assertEquals(original, new UltiStudent(retrieved));
    //    }

    @Test
    public void getAddressBookFilePath() {
        assertNotNull(storageManager.getUltiStudentFilePath());
    }

}
